requirements.txt
aenum==3.0.0
aiohttp==3.7.3
appdirs==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.6
distlib==0.3.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-knox==4.1.0
django-seed==0.2.2
djangorestframework==3.12.1
ecdsa==0.14.1
Faker==4.14.0
filelock==3.0.12
flake8==3.8.4
flatdict==4.0.1
idna==2.10
iniconfig==1.1.1
mccabe==0.6.1
multidict==5.1.0
okta==1.4.0
packaging==20.9
pipenv==2018.11.26
pluggy==0.13.1
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
pydash==4.9.2
pyfakefs==4.3.3
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-mock==3.5.1
pytest-recording==0.11.0
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2020.1
PyYAML==5.4.1
requests==2.24.0
rsa==4.7
six==1.14.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.11
vcrpy==4.1.1
views==0.3
virtualenv==20.0.16
virtualenv-clone==0.5.4
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.6.3

pip~=20.2.4
setuptools~=50.3.2
codegen~=1.0